modules/location_consumer_microservice/requirements.txt
psycopg2-binary==2.9.5
grpcio==1.47.2
grpcio-tools==1.47.2
kafka-python==2.0.2